+++ src/App.tsx
@@ -1,14 +1,14 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Layout from "@/shared/Layout.tsx";
 import GlobalStyle from "@/shared/styles/GlobalStyle.tsx";
-import Home from "@/pages/home/Home.tsx";
-import Search from "@/pages/search/Search.tsx";
+import Home from "@/pages/homeAndSearch/Home";
 import Guide from "@/pages/guide/Guide.tsx";
 import Profile from "@/pages/profile/Profile.tsx";
 import ProtectedRoute from "@/shared/component/ProtectedRoute.tsx";
 import Login from "@/pages/auth/Login.tsx";
 import Signup from "@/pages/auth/Signup";
 import Create from "@/pages/playlist/Create";
+import Search from "./pages/homeAndSearch/Search";
 // import Storage from "@/pages/storage/Storage";
 
 const router = createBrowserRouter([
