--- src/pages/search/Search.tsx
@@ -1,5 +1,0 @@
-const Search = () => {
-  return <h1>Search</h1>;
-};
-
-export default Search;
